refactor(gateway): split internal source reply helpers
